requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
idna==2.10
img2pdf==0.4.1
lxml==4.6.3
numpy==1.20.3
pikepdf==2.12.2
Pillow==8.2.0
pycparser==2.20
pyvips==2.1.14
requests==2.25.1
soupsieve==2.2.1
urllib3==1.26.5
